read and save data user

+++ app/src/main/java/me/creese/sport/util/UserData.java
@@ -3,7 +3,7 @@ package me.creese.sport.util;
 public class UserData {
     public static long BIRTH_DATE = 0;
     public static int SEX;
-    public static float HEIGHT;
-    public static float WEIGHT;
+    public static double HEIGHT;
+    public static double WEIGHT;
 
 }
